requirements.txt
streamlit==1.10.0
numpy==1.19.5
opencv-python-headless==4.6.0.66
pandas==1.1.5
Pillow==8.3.1
tensorflow==2.7.0
keras==2.7.0